Implement metaclasses.
This was a little trickier than I expected, because I completely failed to
understand Smalltalk's model (I long since imbibed ObjVLisp's model, which is
simpler, more uniform -- and much different!). lang_test/metaclasses.som is
probably the easiest way to get some sense of what's going on (and this produces
exactly the same output as Java SOM!). In essence if you have a SOM file such
as:

  C = (
    |x|
    m1 = ( 'm1' println. )
    ------
    |y|
    m2 = ( 'm2' println. )
  )

you're actually defining two classes. The class C is:

  C = (
    |x|
    m1 = ( 'm1' println. )
  )

and its metaclass:

  Cclass = (
    |y|
    m2 = ( 'm2' println. )
  )

C is an instance of Cclass and Cclass is an instance of Metaclass.

The challenge here is that the metaclass model means that when we're
initialising the VM, some objects are referenced before they're created. That
leaves `Val::ilegal`s lurking around, which, if dereferenced, cause bad things
to happen (segfaults and possibly worse). We thus have to patch some objects
after they're created, which requires us to put a couple of `UnsafeCell`s in
`Class` so that we can update them after object creation. This is a bit fiddly,
but also localised enough that it's easy to reasons about the `UnsafeCell`s.

Notice that class fields do not work properly yet, because I did not realise
that Classes can have instance fields. Rather than make this commit even bigger,
I think that's best left to another commit.