Implement class fields.
My original assumption was that only instances of Instance had instance fields,
but it turns out that classes can as well. We thus have to move the
inst_var_lookup and inst_var_set methods into the Obj trait so that anything can
implement them. That means shuffling the relevant code in Instance and
appropriate adjustments to the core interpreter. Fortunately none of this is
particularly complex.